minor bug fixes

+++ src/main.cpp
@@ -46,7 +46,7 @@ int main () {
         glDrawElements(GL_TRIANGLES, 3, GL_UNSIGNED_INT, 0);
 
         shader_program_blue.use();
-        glDrawElements(GL_TRIANGLES, 3, GL_UNSIGNED_INT, (void*)(sizeof(float) * 3));
+        glDrawElements(GL_TRIANGLES, 3, GL_UNSIGNED_INT, (void*)(sizeof(uint32_t) * 3));
 
         window.poll_events();
         window.swap_buffers();
